+++ cups/http-addr.c
@@ -181,7 +181,7 @@ httpAddrListen(http_addr_t *addr,	/* I - Address to bind to */
 
   if ((fd = socket(addr->addr.sa_family, SOCK_STREAM, 0)) < 0)
   {
-    _cupsSetHTTPError(HTTP_STATUS_ERROR);
+    _cupsSetError(IPP_STATUS_ERROR_INTERNAL, strerror(errno), 0);
     return (-1);
   }
 
@@ -240,7 +240,7 @@ httpAddrListen(http_addr_t *addr,	/* I - Address to bind to */
 
   if (status)
   {
-    _cupsSetHTTPError(HTTP_STATUS_ERROR);
+    _cupsSetError(IPP_STATUS_ERROR_INTERNAL, strerror(errno), 0);
 
     close(fd);
 
@@ -253,7 +253,7 @@ httpAddrListen(http_addr_t *addr,	/* I - Address to bind to */
 
   if (listen(fd, INT_MAX))
   {
-    _cupsSetHTTPError(HTTP_STATUS_ERROR);
+    _cupsSetError(IPP_STATUS_ERROR_INTERNAL, strerror(errno), 0);
 
     close(fd);